Ex. 4dot15-4dot23

+++ part_4/routes/blogs.js
@@ -1,48 +1,75 @@
-const express = require('express')
-const app = express()
-const { Blog } = require('../controllers')
+const blogsRouter = require('express').Router()
+const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
-app.get('/', (request, response) => {
-  Blog.find({})
-    .then(blogs => {
-      response.json(blogs)
-    })
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   }
+//   return null
+// }
+
+blogsRouter.get('/', async (request, response) => {
+  const blogs = await Blog.find({}).populate('user')
+  response.json(blogs)
 })
 
-app.post('/', (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const { title, author, url, likes } = request.body
   if (!title || !url) {
     return response.status(400).json({ error: 'Title and URL are required' })
   }
-
+  //const decodedToken = jwt.verify(getTokenFrom(request), process.env.SEKRIT)
+  // const decodedToken = jwt.verify(request.token, process.env.SEKRIT)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  const user = request.user
+  //const user = await User.findById(decodedToken.id)
+  //const randUser = await User.findOne()
   const blog = new Blog({
     title : title,
     author : author,
     url : url,
-    likes: likes
+    likes: likes,
+    user: user.id
   })
-
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
-app.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
   const { id } = request.params
-  
+  // if (!request.token) {
+  //   return response.status(401).json({ error: 'token missing' })
+  // }
+  // const decodedToken = jwt.verify(request.token, process.env.SEKRIT)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  const user = request.user
+  const blog = await Blog.findById(id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'Only the creator of a blog can delete it' })
+  }
   try {
-    const deletedBlog = await Blog.findByIdAndDelete(id)
+    await Blog.findByIdAndDelete(id)
     response.status(204).end()
   } catch (error) {
     response.status(500).json({ error: 'Failed to delete blog' })
   }
 })
 
-app.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', async (request, response) => {
   const {likes} = request.body
-
   try {
     const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, { likes: likes }, { new: true, runValidators: true })
     if (updatedBlog) {
@@ -56,4 +83,4 @@ app.put('/:id', async (request, response) => {
   }
 })
 
-module.exports = app
+module.exports = blogsRouter
